requirements.txt
Django
djangorestframework
pandas
matplotlib
reportlab
python-docx
openpyxl
